pyproject.toml
[project]
name = "microbackdjango-user"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "uvicorn (>=0.34.1,<0.35.0)",
    "django (>=5.2,<6.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "dj-database-url (>=2.3.0,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pathlib (>=1.0.1,<2.0.0)",
    "rest-framework-simplejwt (>=0.0.2,<0.0.3)",
    "whitenoise (>=6.9.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
